chore(customlogger): remove unnecessary conditional branches
`Infof()` and `Errorf()` for now at least, is not going to care
about the case if `CustomLogger.Info` or `CustomLogger.Error` is
not nil but `CustomLogger.Logfile.File` is nil. User has to
initialize the `CustomLogger` struct with `CustomLoggerInit()`
to avoid any unnecessary problems caused by struct members not
being initialized with nil.